Refactor from self-review
